+++ packages/iov-ethereum/src/index.ts
@@ -1,7 +1,8 @@
 // high level exports
+export { Erc20Options } from "./erc20";
 export { EthereumConnection, EthereumConnectionOptions } from "./ethereumconnection";
 export { ethereumConnector } from "./ethereumconnector";
-export { ethereumCodec } from "./ethereumcodec";
+export { ethereumCodec, EthereumCodec, EthereumCodecOptions } from "./ethereumcodec";
 
 // Custom helper functions
 export { pubkeyToAddress, toChecksummedAddress } from "./address";