+++ packages/iov-ethereum/types/index.d.ts
@@ -1,4 +1,5 @@
+export { Erc20Options } from "./erc20";
 export { EthereumConnection, EthereumConnectionOptions } from "./ethereumconnection";
 export { ethereumConnector } from "./ethereumconnector";
-export { ethereumCodec } from "./ethereumcodec";
+export { ethereumCodec, EthereumCodec, EthereumCodecOptions } from "./ethereumcodec";
 export { pubkeyToAddress, toChecksummedAddress } from "./address";